Fix MOVED and ASK codepaths
The MOVED and ASK codepaths were broken whe a new node joins the cluster
and starts having data. The code issue was trying to insert the new node
from user process instead of the ecredis_server gen server.

The new logic is like this:
1. If we get a MOVED or ASK result, First check if there is already
connection to the new node.
2. If we are already connected, just execute the query again on the
right node, and if the move is permanent notify the ecredis_server
process to update the mapping for this slow and schedule cluster remap.
3. If there is no connection to the new node, ask ecredis_server to
add this node first and then, follow the same logic as 2.

While working on this I notices few other potential issues with the code
for which I mostly left TODOs, or did small fixes if they were easy.

I also added test for this new use case that has number of helper
functions to add a new node and migrate some data to it, then migrate
the data back and remove it from the cluster.