fix qp asking
If we do qp query and some of our queries receive asking replies we were
not handling this properly. One issue is that when a command fails with
ASK error, you have to generate 2 commands [["ASKING"],
OriginalCommand],

This is a problem when we try to merge the error messages and re-execute
the commands since some things might return a single command and some
places might return this list of commands.

Update test/moved_tests.erl

Co-authored-by: vipin211 <vipin211@hotmail.com>

Fixes